Use configurable import paths for todo models
